+++ js/create-descriptions.js
@@ -1,13 +1,7 @@
-import { getNumInRange } from "./get-num-in-range";
-import { getUniqueId } from "./get-unique-id";
-import { getRandomElementIn } from "./get-random-element";
+import { getUniqueId, getNumInRange, getRandomElementIn } from './util.js';
 
 // Массивы
-const COMMENTSTEXT = ['Всё отлично!','В целом всё неплохо. Но не всё.',
-'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?'];
+const COMMENTSTEXT = ['Всё отлично!','В целом всё неплохо. Но не всё.','Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.','Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.','Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.','Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?'];
 const NAMES = ['Ваня', 'Егор', 'Аля', 'Евдоким'];
 const DESCRIPTIONS = ['Вот такая замечательная фоточка)', 'Я на море!!', 'Вот это еда!!', 'Холодильник'];
