--- js/get-num-in-range.js
@@ -1,9 +1,0 @@
-// Функция генерации случайных чисел в диапазоне
-const getNumInRange = (min, max) => {
-  const minNum = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const maxNum = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.floor(Math.random() * (maxNum - minNum + 1) + minNum);
-  return result;
-};
-
-export {getNumInRange};
