--- js/get-unique-id.js
@@ -1,9 +1,0 @@
-// Функция получения id
-const getUniqueId = () => {
-  let id = 0;
-  return function () {
-    id += 1;
-    return id;
-  };
-};
-export {getUniqueId};
