+++ js/main.js
@@ -1,4 +1,5 @@
-import { getPhotoDescription } from "./create-description-for-photo";
+import { getPhotoDescription } from './create-descriptions.js';
 
 // генерируем 25 описаний для фото
+// eslint-disable-next-line no-unused-vars
 const photoDescriptions = Array.from({length: 25}, getPhotoDescription);
